+++ src/API/exports.jl
@@ -103,11 +103,6 @@ export pyxor
 export @pyconvert
 export pyconvert
 export pyconvert_add_rule
-export PYCONVERT_PRIORITY_ARRAY
-export PYCONVERT_PRIORITY_CANONICAL
-export PYCONVERT_PRIORITY_FALLBACK
-export PYCONVERT_PRIORITY_NORMAL
-export PYCONVERT_PRIORITY_WRAP
 export pyconvert_return
 export pyconvert_unconverted