+++ src/Convert/pandas.jl
@@ -3,10 +3,10 @@ pyconvert_rule_pandas_na(::Type{Missing}, x::Py) = pyconvert_return(missing)
 
 function init_pandas()
     pyconvert_add_rule(
+        pyconvert_rule_pandas_na,
         "pandas._libs.missing:NAType",
         Missing,
-        pyconvert_rule_pandas_na,
-        PYCONVERT_PRIORITY_CANONICAL,
+        Any,
     )
-    pyconvert_add_rule("pandas._libs.missing:NAType", Nothing, pyconvert_rule_pandas_na)
+    pyconvert_add_rule(pyconvert_rule_pandas_na, "pandas._libs.missing:NAType", Nothing, Nothing)
 end